backend/requirements.txt
# Core Framework
fastapi==0.131.0
uvicorn[standard]==0.41.0
python-multipart==0.0.22
websockets==16.0

# LangChain & LangGraph (Required)
langchain==1.2.10
langchain-openai==1.1.10
langgraph==1.0.9
langgraph-checkpoint-postgres==3.0.4

# Observability (Required)
langfuse==3.14.5

# Database & ORM
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
alembic==1.18.4
psycopg2-binary==2.9.11

# Cache
redis==7.2.0
hiredis==3.3.0

# Voice Integration
deepgram-sdk==6.0.0
elevenlabs==2.36.1

# Security & Auth
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.12.5

# Utilities
python-dotenv==1.2.1
pydantic-settings==2.13.1
httpx==0.28.1
tenacity==9.1.4

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1

# Type Checking
mypy==1.19.1
types-redis==4.6.0.20241004

# Development
black==26.1.0
ruff==0.15.2
faker==40.4.0
